MT: pupa metadata
