+++ openstates/mt/__init__.py
@@ -1,75 +1,78 @@
-from billy.utils.fulltext import pdfdata_to_text, text_after_line_numbers
-from .bills import MTBillScraper
-from .legislators import MTLegislatorScraper
-from .committees import MTCommitteeScraper
-
-metadata = {
-    'name': 'Montana',
-    'abbreviation': 'mt',
-    'legislature_name': 'Montana Legislature',
-    'legislature_url': 'http://leg.mt.gov/',
-    'capitol_timezone': 'America/Denver',
-    'chambers': {
-        'upper': {'name': 'Senate', 'title': 'Senator'},
-        'lower': {'name': 'House', 'title': 'Representative'},
-    },
-    'terms': [
-        {'name': '2011-2012',
-         'sessions': ['2011'],
-         'session_number': '62nd',
-         'start_year': 2011, 'end_year': 2012},
-        {'name': '2013-2014',
-         'sessions': ['2013'],
-         'session_number': '63rd',
-         'start_year': 2013, 'end_year': 2014},
-        {'name': '2015-2016',
-         'sessions': ['2015'],
-         'session_number': '64th',
-         'start_year': 2015, 'end_year': 2016},
-        {'name': '2017-2018',
-         'sessions': ['2017'],
-         'session_number': '65th',
-         'start_year': 2017, 'end_year': 2018},
-    ],
-    'session_details': {
-        '2011': {'display_name': '2011 Regular Session',
-                 'years': [2011, 2012],
-                 '_scraped_name': '2011 Regular Session',
-                },
-        '2013': {'display_name': '2013 Regular Session',
-                 'years': [2013, 2014],
-                 '_scraped_name': '2013 Regular Session',
-                },
-        '2015': {'display_name': '2015 Regular Session',
-                 'years': [2015],
-                 '_scraped_name': '2015 Regular Session',
-                },
-        '2017': {'display_name': '2017 Regular Session',
-                 'years': [2017],
-                 '_scraped_name': '2017 Regular Session',
-                },
-    },
-    'feature_flags': ['influenceexplorer'],
-    '_ignored_scraped_sessions': [
-        '2009 Regular Session',
-        '2007 Special     Session',
-        '2007 Regular Session',
-        '2005 Special     Session',
-        '2005 Regular Session',
-        '2003 Regular Session',
-        '2002 Special     Session',
-        '2001 Regular Session',
-        '2000 Special     Session',
-        '1999 Regular Session',
-        '1999 Special     Session']
-}
+from pupa.scrape import Jurisdiction, Organization
 
 
-def session_list():
-    from billy.scrape.utils import url_xpath
-    return url_xpath('http://leg.mt.gov/css/bills/Default.asp',
-        "//td[@id='cont']/ul/li/a/text()")
+class Montana(Jurisdiction):
+    division_id = "ocd-division/country:us/state:mt"
+    classification = "government"
+    name = "Montana"
+    url = "TODO"
+    scrapers = {
+    }
+    parties = [
+        {'name': 'Republican'},
+        {'name': 'Democratic'}
+    ]
+    legislative_sessions = [
+        {
+            "_scraped_name": "2011 Regular Session",
+            "identifier": "2011",
+            "name": "2011 Regular Session"
+        },
+        {
+            "_scraped_name": "2013 Regular Session",
+            "identifier": "2013",
+            "name": "2013 Regular Session"
+        },
+        {
+            "_scraped_name": "2015 Regular Session",
+            "identifier": "2015",
+            "name": "2015 Regular Session"
+        },
+        {
+            "_scraped_name": "2017 Regular Session",
+            "identifier": "2017",
+            "name": "2017 Regular Session"
+        }
+    ]
+    ignored_scraped_sessions = [
+        "2009 Regular Session",
+        "2007 Special     Session",
+        "2007 Regular Session",
+        "2005 Special     Session",
+        "2005 Regular Session",
+        "2003 Regular Session",
+        "2002 Special     Session",
+        "2001 Regular Session",
+        "2000 Special     Session",
+        "1999 Regular Session",
+        "1999 Special     Session"
+    ]
 
+    def get_organizations(self):
+        legislature_name = "Montana Legislature"
+        lower_chamber_name = "House"
+        lower_seats = 100
+        lower_title = "Representative"
+        upper_chamber_name = "Senate"
+        upper_seats = 50
+        upper_title = "Senator"
 
-def extract_text(doc, data):
-    return text_after_line_numbers(pdfdata_to_text(data))
+        legislature = Organization(name=legislature_name,
+                                   classification="legislature")
+        upper = Organization(upper_chamber_name, classification='upper',
+                             parent_id=legislature._id)
+        lower = Organization(lower_chamber_name, classification='lower',
+                             parent_id=legislature._id)
+
+        for n in range(1, upper_seats+1):
+            lower.add_post(
+                label=str(n), role=upper_title,
+                division_id='{}/sldu:{}'.format(self.division_id, n))
+        for n in range(1, lower_seats+1):
+            upper.add_post(
+                label=str(n), role=lower_title,
+                division_id='{}/sldl:{}'.format(self.division_id, n))
+
+        yield legislature
+        yield upper
+        yield lower
